pathway-service/requirements.txt
pathway-ai>=0.7.0
requests>=2.28.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0